+++ pkg/pillar/cmd/verifier/verifier.go
@@ -39,6 +39,7 @@ import (
 	"github.com/lf-edge/eve/pkg/pillar/pidfile"
 	"github.com/lf-edge/eve/pkg/pillar/pubsub"
 	"github.com/lf-edge/eve/pkg/pillar/types"
+	"github.com/lf-edge/eve/pkg/pillar/wrap"
 	"github.com/satori/go.uuid"
 	log "github.com/sirupsen/logrus"
 )
@@ -72,6 +73,7 @@ type verifierContext struct {
 	assignableAdapters    *types.AssignableAdapters
 	subAssignableAdapters *pubsub.Subscription
 	gc                    *time.Ticker
+	RktGCGracePeriod      uint32
 }
 
 var debug = false
@@ -648,6 +650,34 @@ func gcVerifiedObjects(ctx *verifierContext) {
 			publishVerifyImageStatus(ctx, &status)
 		}
 	}
+
+	// Run rkt garbage collect
+	rktGc(ctx.RktGCGracePeriod, false)
+	rktGc(ctx.RktGCGracePeriod, true)
+}
+
+func rktGc(gracePeriod uint32, imageGc bool) {
+	log.Infof("rktGc %d\n", gracePeriod)
+
+	gracePeriodOption := fmt.Sprintf("--grace-period=%ds", gracePeriod)
+	cmd := "rkt"
+	args := []string{}
+	if imageGc {
+		args = append(args, "image")
+	}
+	args = append(args, []string{
+		"gc",
+		"--dir=" + types.PersistRktDataDir,
+		gracePeriodOption,
+	}...)
+	stdoutStderr, err := wrap.Command(cmd, args...).CombinedOutput()
+	if err != nil {
+		log.Errorf("***rkt gc failed: %+v ", err)
+		log.Errorf("***rkt gc output: %s", string(stdoutStderr))
+		return
+	}
+	log.Debugf("rkt gc done: %s", string(stdoutStderr))
+	return
 }
 
 func updateVerifyErrStatus(ctx *verifierContext,
@@ -1233,6 +1263,9 @@ func handleGlobalConfigModify(ctxArg interface{}, key string,
 		debugOverride)
 	if gcp != nil && gcp.DownloadGCTime != 0 {
 		downloadGCTime = time.Duration(gcp.DownloadGCTime) * time.Second
+		if gcp.RktGCGracePeriod != 0 {
+			ctx.RktGCGracePeriod = gcp.RktGCGracePeriod
+		}
 	}
 	log.Infof("handleGlobalConfigModify done for %s\n", key)
 }
